Issue 439: merge config and message models.
