+++ app/src/main/resources/struts.xml
@@ -215,9 +215,9 @@
             <result name="list" type="tiles">.PlanetEdit</result>
         </action>
         
-        <action name="planets!*" method="{1}"
+        <action name="planets"
                 class="org.apache.roller.weblogger.ui.struts2.admin.Planets">
-            <result name="list" type="tiles">.Planets</result>
+            <result name="success" type="tiles">.Planets</result>
         </action>
     </package>
