+++ app/src/main/webapp/WEB-INF/velocity/templates/error-page.vm
@@ -18,7 +18,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <title>$text.get("errorPage.title")</title>
+    <title>$server.msg("errorPage.title")</title>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <style>
@@ -47,7 +47,7 @@
             <tr>
                 <td class="bannerLeft"></td>
                 <td class="bannerRight">
-                   <a href="$url.absoluteSite">$config.siteName</a>
+                   <a href="$url.absoluteSite">$server.siteName</a>
                 </td>
             </tr>
         </table>
@@ -58,8 +58,8 @@
       <h2 class="error">Velocity Template Error</h2>
       <span class="error">
         <ul>
-          <li>$text.get("errorPage.title")</li>
-          <li>$text.get("errorPage.message")</li>
+          <li>$server.msg("errorPage.title")</li>
+          <li>$server.msg("errorPage.message")</li>
         </ul>
       </span>
       <table width="80%" class="dottedtable" style="border-collapse: collapse;">
