+++ app/src/main/webapp/WEB-INF/velocity/templates/feeds/comments-atom.vm
@@ -26,10 +26,10 @@
     <link rel="alternate" type="text/html" href="$url.home" />
   #end
     <link rel="self" type="application/atom+xml" href="$url.feed.comments.atom" />
-    #if($config.feedHistoryEnabled)
+    #if($server.feedHistoryEnabled)
         #showFeedHistory($pager)
     #end
     <updated>$utils.formatIso8601Date($pager.lastUpdated)</updated>
-    <generator uri="https://github.com/gmazza/tightblog" version="$config.systemVersion">TightBlog</generator>
+    <generator uri="https://github.com/gmazza/tightblog" version="$server.systemVersion">TightBlog</generator>
     #showCommentsAtom10($pager.items)
 </feed>
