+++ app/src/main/webapp/WEB-INF/velocity/templates/feeds/entries-atom.vm
@@ -36,10 +36,10 @@
     #else
         <link rel="self" type="application/atom+xml" href="$url.feed.entries.atom" />
     #end
-    #if($config.feedHistoryEnabled)
+    #if($server.feedHistoryEnabled)
         #showFeedHistory($pager)
     #end
     <updated>$utils.formatIso8601Date($pager.lastUpdated)</updated>
-    <generator uri="https://github.com/gmazza/tightblog" version="$config.systemVersion">TightBlog</generator>
+    <generator uri="https://github.com/gmazza/tightblog" version="$server.systemVersion">TightBlog</generator>
     #showEntriesAtom10($pager.items)
 </feed>
