+++ app/src/main/webapp/themes/basic/_day.vm
@@ -4,7 +4,7 @@
 <div class="dayBox">
 
     <div class="dayTitle">
-       $utils.formatDate($day, $text.get("macro.weblog.date.toStringFormat"))
+       $utils.formatDate($day, $server.msg("macro.weblog.date.toStringFormat"))
     </div>
 
     #foreach( $entry in $entries )
@@ -19,15 +19,15 @@
             #end
         </p>
         <p class="entryInfo">
-            Posted at <a href="$url.entry($entry.anchor)">$utils.formatDate($entry.pubTime, $text.get("macro.weblog.datetime.toStringFormat"))</a>
+            Posted at <a href="$url.entry($entry.anchor)">$utils.formatDate($entry.pubTime, $server.msg("macro.weblog.datetime.toStringFormat"))</a>
             by $entry.creator.screenName in <span class="category">$entry.category.name</span> &nbsp;|&nbsp;
             #if ($model.isUserBlogPublisher($entry.weblog))
-                <a href="$url.editEntry($entry.anchor)">$text.get("generic.edit")</a> &nbsp;|&nbsp;
+                <a href="$url.editEntry($entry.anchor)">$server.msg("generic.edit")</a> &nbsp;|&nbsp;
             #end
             #set($commentCount = $entry.commentCount)
             #if($entry.commentsStillAllowed || $commentCount > 0)
                 #set($link = "$url.comments($entry.anchor)" )
-                <a href="$link" class="commentsLink">$text.get("macro.weblog.comments") [$commentCount]</a>
+                <a href="$link" class="commentsLink">$server.msg("macro.weblog.comments") [$commentCount]</a>
             #end
         </p>
         #showTrackbackAutodiscovery($entry)
