+++ app/src/main/webapp/themes/basicmobile/_day-mobile.vm
@@ -2,7 +2,7 @@
 ## $entry here is a WeblogEntry
 <!-- _day-mobile -->
 <ul data-role="listview" data-divider-theme="b" data-inset="true">
-  <li data-role="list-divider">$utils.formatDate($day, $text.get("macro.weblog.date.toStringFormat"))</li>
+  <li data-role="list-divider">$utils.formatDate($day, $server.msg("macro.weblog.date.toStringFormat"))</li>
   ## Loop entries
   #foreach( $entry in $entries )
     <li data-role="list-divider" data-theme="a">
@@ -10,10 +10,10 @@
          $entry.title
          #set($commentCount = $entry.commentCount)
          #if($model.isUserBlogPublisher($entry.weblog))
-           &nbsp;|&nbsp;&nbsp;<a href="$url.editEntry($entry.anchor)" data-role="button" data-inline="true" data-mini="true">$text.get("generic.edit")</a>
+           &nbsp;|&nbsp;&nbsp;<a href="$url.editEntry($entry.anchor)" data-role="button" data-inline="true" data-mini="true">$server.msg("generic.edit")</a>
          #end
          #if($entry.commentsStillAllowed || $commentCount > 0)
-           |&nbsp;&nbsp;$text.get("macro.weblog.comments") [$commentCount]
+           |&nbsp;&nbsp;$server.msg("macro.weblog.comments") [$commentCount]
          #end
        </div>
     </li>
@@ -34,7 +34,7 @@
           </p>
         #end
         <p class="ui-li-desc content-textbox">
-          Posted at $utils.formatDate($entry.pubTime, $text.get("macro.weblog.datetime.toStringFormat"))
+          Posted at $utils.formatDate($entry.pubTime, $server.msg("macro.weblog.datetime.toStringFormat"))
           by $entry.creator.screenName in <span class="category">$entry.category.name</span>
         </p>
       #if(!$model.permalink)
