+++ app/src/main/webapp/themes/basicmobile/weblog-mobile.vm
@@ -63,10 +63,10 @@
       #else
         ## No entries.
         <ul data-role="listview" data-divider-theme="b" data-inset="true">
-          <li data-role="list-divider">$utils.formatDate($utils.now, $text.get("macro.weblog.date.toStringFormat"))</li>
+          <li data-role="list-divider">$utils.formatDate($utils.now, $server.msg("macro.weblog.date.toStringFormat"))</li>
           <li data-theme="a">
             <span class="ui-li-desc content-textbox">
-              <h3>$text.get("macro.weblog.noEntriesForCategory")</h3>
+              <h3>$server.msg("macro.weblog.noEntriesForCategory")</h3>
             </span>
           </li>
         </ul>
