+++ app/src/main/webapp/themes/fauxcoly/archives.vm
@@ -112,7 +112,7 @@
     <script>
         function validateSearch(form) {
             if (form.q.value == "") {
-                alert("$text.get("macro.weblog.searchalert")");
+                alert("$server.msg("macro.weblog.searchalert")");
                 form.q.focus();
                 return false;
             }
