+++ app/src/main/webapp/themes/fauxcoly/day.vm
@@ -1,6 +1,6 @@
 #if(!$model.permalink)
 <div class="daybox">
-    <h4 class="dayTitle">$utils.formatDate($day, $text.get("macro.weblog.date.toStringFormat"))</h4>
+    <h4 class="dayTitle">$utils.formatDate($day, $server.msg("macro.weblog.date.toStringFormat"))</h4>
 #end
 
     #foreach( $entry in $entries )
@@ -25,18 +25,18 @@
             $entry.creator.screenName in <span class="category">$entry.category.name</span>
 
             <a class="permalink" href="$url.entry($entry.anchor)">
-                $utils.formatDate($entry.pubTime, $text.get("macro.weblog.datetime.toStringFormat"))</a>
+                $utils.formatDate($entry.pubTime, $server.msg("macro.weblog.datetime.toStringFormat"))</a>
 
             #set($commentCount = $entry.commentCount)
             #if($entry.commentsStillAllowed || $commentCount > 0)
                 #set($link = "$url.comments($entry.anchor)" )
-                <a href="$link" class="commentslink">$text.get("macro.weblog.comments") [$commentCount]</a>
+                <a href="$link" class="commentslink">$server.msg("macro.weblog.comments") [$commentCount]</a>
             #end
 
             #if ($model.isUserBlogPublisher($entry.weblog))
                 &nbsp;|&nbsp;
                 <a href="$url.editEntry($entry.anchor)">
-                    $text.get("generic.edit")</a>
+                    $server.msg("generic.edit")</a>
             #end
         </p>
