+++ app/src/main/webapp/themes/fauxcoly/weblog.vm
@@ -68,7 +68,7 @@
     <script>
         function validateSearch(form) {
             if (form.q.value == "") {
-                alert("$text.get("macro.weblog.searchalert")");
+                alert("$server.msg("macro.weblog.searchalert")");
                 form.q.focus();
                 return false;
             }
