+++ app/src/main/webapp/themes/gaurav/day.vm
@@ -15,7 +15,7 @@
 			<p class="lead">by <span style="text-transform: capitalize;">$entry.creator.screenName</span></p>
 		    
 			<hr>
-			<p><i class="icon-time"></i> Posted on $utils.formatDate($day, $text.get("macro.weblog.date.toStringFormat")) at $utils.formatDate($entry.pubTime, $text.get("macro.weblog.time.toStringFormat")) in <span class="category">$entry.category.name</span> </p>
+			<p><i class="icon-time"></i> Posted on $utils.formatDate($day, $server.msg("macro.weblog.date.toStringFormat")) at $utils.formatDate($entry.pubTime, $server.msg("macro.weblog.time.toStringFormat")) in <span class="category">$entry.category.name</span> </p>
 			<hr>
 	
 		    <div class="entryContent">
@@ -35,13 +35,13 @@
 		        #set($commentCount = $entry.commentCount)
 		        #if($entry.commentsStillAllowed || $commentCount > 0)
 		            #set($link = "$url.comments($entry.anchor)" )
-		            <a href="$link" class="commentslink">$text.get("macro.weblog.comments") [$commentCount]</a>
+		            <a href="$link" class="commentslink">$server.msg("macro.weblog.comments") [$commentCount]</a>
 		        #end
 
 		        #if ($model.isUserBlogPublisher($entry.weblog))
 		            &nbsp;|&nbsp;
 		            <a href="$url.editEntry($entry.anchor)">
-		                $text.get("generic.edit")</a>
+		                $server.msg("generic.edit")</a>
 		        #end
 		    </p>
